# openapi/licoreria.yaml
openapi: 3.1.1
info:
  title: API Licorería / Discoteca
  description: >-
    Contrato generado desde el código de la API del sistema de licorería y discoteca (sucursal única). Los errores se
    devuelven conforme a RFC 7807.
  version: v1
paths:
...
  /api/v1/variantes/{id}/recetas:
    get:
      tags:
        - Variantes
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RecetaDto'
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RecetaDto'
            text/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RecetaDto'
    post:
      tags:
        - Variantes
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecetaCrearDto'
          text/json:
            schema:
              $ref: '#/components/schemas/RecetaCrearDto'
          application/*+json:
            schema:
              $ref: '#/components/schemas/RecetaCrearDto'
        required: true
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/RecetaDto'
            application/json:
              schema:
                $ref: '#/components/schemas/RecetaDto'
            text/json:
              schema:
                $ref: '#/components/schemas/RecetaDto'
  /api/v1/variantes/{id}/recetas/{recetaId}:
    delete:
      tags:
        - Variantes
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: recetaId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
...
  /api/v1/zonas:
    get:
      tags:
        - Zonas
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ZonaDto'
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ZonaDto'
            text/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ZonaDto'
    post:
      tags:
        - Zonas
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ZonaCrearDto'
          text/json:
            schema:
              $ref: '#/components/schemas/ZonaCrearDto'
          application/*+json:
            schema:
              $ref: '#/components/schemas/ZonaCrearDto'
        required: true
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ZonaDto'
            application/json:
              schema:
                $ref: '#/components/schemas/ZonaDto'
            text/json:
              schema:
                $ref: '#/components/schemas/ZonaDto'
  /api/v1/zonas/{id}:
    put:
      tags:
        - Zonas
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ZonaEditarDto'
          text/json:
            schema:
              $ref: '#/components/schemas/ZonaEditarDto'
          application/*+json:
            schema:
              $ref: '#/components/schemas/ZonaEditarDto'
        required: true
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ZonaDto'
            application/json:
              schema:
                $ref: '#/components/schemas/ZonaDto'
            text/json:
              schema:
                $ref: '#/components/schemas/ZonaDto'
    delete:
      tags:
        - Zonas
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
components:
  schemas:
...
    RecetaCrearDto:
      required:
        - varianteInsumoId
        - cantidad
      type: object
      properties:
        varianteInsumoId:
          type: string
          format: uuid
        cantidad:
          pattern: ^-?(?:0|[1-9]\d*)(?:\.\d+)?$
          type:
            - number
            - string
          format: double
    RecetaDto:
      required:
        - id
        - varianteInsumoId
        - varianteInsumoNombre
        - sku
        - cantidad
      type: object
      properties:
        id:
          type: string
          format: uuid
        varianteInsumoId:
          type: string
          format: uuid
        varianteInsumoNombre:
          type: string
        sku:
          type: string
        cantidad:
          pattern: ^-?(?:0|[1-9]\d*)(?:\.\d+)?$
          type:
            - number
            - string
          format: double
...
    TipoZona:
      type: integer
...
    ZonaCrearDto:
      required:
        - nombre
        - tipo
      type: object
      properties:
        nombre:
          type: string
        tipo:
          $ref: '#/components/schemas/TipoZona'
        activo:
          type: boolean
          default: true
        color:
          type:
            - 'null'
            - string
        posX:
          pattern: ^-?(?:0|[1-9]\d*)(?:\.\d+)?$
          type:
            - number
            - string
          format: double
          default: 0
        posY:
          pattern: ^-?(?:0|[1-9]\d*)(?:\.\d+)?$
          type:
            - number
            - string
          format: double
          default: 0
        ancho:
          pattern: ^-?(?:0|[1-9]\d*)(?:\.\d+)?$
          type:
            - number
            - string
          format: double
          default: 0
        alto:
          pattern: ^-?(?:0|[1-9]\d*)(?:\.\d+)?$
          type:
            - number
            - string
          format: double
          default: 0
    ZonaDto:
      required:
        - id
        - nombre
        - tipo
        - activo
        - color
        - posX
        - posY
        - ancho
        - alto
      type: object
      properties:
        id:
          type: string
          format: uuid
        nombre:
          type: string
        tipo:
          $ref: '#/components/schemas/TipoZona'
        activo:
          type: boolean
        color:
          type:
            - 'null'
            - string
        posX:
          pattern: ^-?(?:0|[1-9]\d*)(?:\.\d+)?$
          type:
            - number
            - string
          format: double
        posY:
          pattern: ^-?(?:0|[1-9]\d*)(?:\.\d+)?$
          type:
            - number
            - string
          format: double
        ancho:
          pattern: ^-?(?:0|[1-9]\d*)(?:\.\d+)?$
          type:
            - number
            - string
          format: double
        alto:
          pattern: ^-?(?:0|[1-9]\d*)(?:\.\d+)?$
          type:
            - number
            - string
          format: double
    ZonaEditarDto:
      required:
        - id
        - nombre
        - tipo
        - activo
      type: object
      properties:
        id:
          type: string
          format: uuid
        nombre:
          type: string
        tipo:
          $ref: '#/components/schemas/TipoZona'
        activo:
          type: boolean
        color:
          type:
            - 'null'
            - string
        posX:
          pattern: ^-?(?:0|[1-9]\d*)(?:\.\d+)?$
          type:
            - number
            - string
          format: double
          default: 0
        posY:
          pattern: ^-?(?:0|[1-9]\d*)(?:\.\d+)?$
          type:
            - number
            - string
          format: double
          default: 0
        ancho:
          pattern: ^-?(?:0|[1-9]\d*)(?:\.\d+)?$
          type:
            - number
            - string
          format: double
          default: 0
        alto:
          pattern: ^-?(?:0|[1-9]\d*)(?:\.\d+)?$
          type:
            - number
            - string
          format: double
          default: 0
  securitySchemes:
    Bearer:
      type: http
      description: 'Ingrese el token JWT. Ejemplo: Bearer {token}'
      scheme: bearer
      bearerFormat: JWT
security:
  - {}
...